# backend/docs_design/openapi.yaml
...
openapi: 3.1.0
info:
  title: BoMesh API
  version: 1.0.0
  description: Contract-first target for workspace knowledge, IAM, integrations, and governed ingestion.
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/connections/{connection_id}/sources:
    post:
      tags: [sources]
      operationId: createConnectionSource
      parameters:
        - {$ref: '#/components/parameters/ConnectionId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SourceCreate'}
      responses:
        '201':
          description: Created Source
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Source'}
  /api/v1/sources:
    get:
      tags: [sources]
      operationId: listSources
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - name: connection_id
          in: query
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: Sources
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SourcePage'}
  /api/v1/sources/{source_id}:
    parameters:
      - {$ref: '#/components/parameters/SourceId'}
    get:
      tags: [sources]
      operationId: getSource
      responses:
        '200':
          description: Source detail
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Source'}
    patch:
      tags: [sources]
      operationId: updateSource
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SourceUpdate'}
      responses:
        '200':
          description: Updated Source
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Source'}
    delete:
      tags: [sources]
      operationId: deleteSource
      responses:
        '204': {description: Source tombstoned}
  /api/v1/sources/{source_id}/syncs:
    post:
      tags: [sources]
      operationId: syncSource
      x-authorization-rule: Same rule as other Source changes; source.manage for workspace connections, the owner for a personal connection. 409 while a sync is running.
      description: >-
        Sync now: discover what changed at the provider, store new or changed
        originals, register or update their Documents (pending when changed) and
        tombstone removed ones. Never processes; processing is an Ingestion Run.
      parameters:
        - {$ref: '#/components/parameters/SourceId'}
      responses:
        '202':
          description: Sync started
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Source'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/sources/{source_id}/schedule:
    parameters:
      - {$ref: '#/components/parameters/SourceId'}
    get:
      tags: [sources]
      operationId: getSourceSchedule
      responses:
        '200':
          description: Source schedule
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Schedule'}
    put:
      tags: [sources]
      operationId: putSourceSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SchedulePut'}
      responses:
        '200':
          description: Replaced schedule
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Schedule'}
    patch:
      tags: [sources]
      operationId: patchSourceSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SchedulePatch'}
      responses:
        '200':
          description: Updated schedule
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Schedule'}
    delete:
      tags: [sources]
      operationId: deleteSourceSchedule
      responses:
        '204': {description: Schedule removed}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PageSize: {name: page_size, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
...
    ConnectionId: {name: connection_id, in: path, required: true, schema: {type: string, format: uuid}}
    SourceId: {name: source_id, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
...
    Forbidden:
      description: Insufficient permission
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    NotFound:
      description: Resource not found or not visible
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Conflict:
      description: Resource or lifecycle conflict
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, request_id, details]
      properties:
        code: {type: string}
        message: {type: string}
        request_id: {type: string, format: uuid}
        details: {type: object, additionalProperties: true}
    PageFields:
      type: object
      required: [page, page_size, total]
      properties:
        page: {type: integer, minimum: 1}
        page_size: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, minimum: 0}
...
    Source:
      type: object
      required: [id, connection_id, collection_id, sync_mode, status]
      properties:
        id: {type: string, format: uuid}
        connection_id: {type: string, format: uuid}
        collection_id: {type: string, format: uuid}
        display_name: {type: [string, 'null']}
        resource_type: {type: [string, 'null']}
        external_resource_id: {type: [string, 'null']}
        sync_mode: {type: string, enum: [manual, scheduled]}
        status: {type: string, enum: [ready, paused, failed, connection_required, disabled]}
        schedule: {anyOf: [{$ref: '#/components/schemas/Schedule'}, {type: 'null'}]}
        sync:
          description: The latest sync (it registers, updates and removes Documents; it never processes). Null before the first sync.
          anyOf:
            - type: object
              required: [status]
              properties:
                status: {type: string, enum: [running, succeeded, failed]}
                last_synced_at: {type: [string, 'null'], format: date-time}
                error: {type: [string, 'null'], description: User-safe reason the sync failed}
                added: {type: integer, minimum: 0}
                updated: {type: integer, minimum: 0}
                removed: {type: integer, minimum: 0}
                failed: {type: integer, minimum: 0}
            - {type: 'null'}
        pending_documents: {type: integer, minimum: 0, description: The Source's Documents that are pending or outdated}
    SourceCreate:
      type: object
      required: [collection_id]
      properties:
        collection_id: {type: string, format: uuid}
        display_name: {type: [string, 'null']}
        resource_type: {type: [string, 'null']}
        external_resource_id: {type: [string, 'null']}
        config: {type: object, additionalProperties: true}
        schedule: {anyOf: [{$ref: '#/components/schemas/SchedulePut'}, {type: 'null'}]}
    SourceUpdate:
      type: object
      additionalProperties: false
      description: Destination collection, connection and provider resource are immutable. Supplied config replaces the connector scope/configuration and resets its checkpoint; omission preserves it.
      minProperties: 1
      properties:
        display_name: {type: string}
        status: {type: string, enum: [ready, paused, disabled]}
        config: {type: object, additionalProperties: true}
    SourcePage:
      allOf:
        - {$ref: '#/components/schemas/PageFields'}
        - type: object
          required: [items]
          properties: {items: {type: array, items: {$ref: '#/components/schemas/Source'}}}
    SchedulePut:
      type: object
      required: [schedule_type, cron_expression, enabled, overlap_policy]
      properties:
        schedule_type: {type: string, enum: [cron, interval]}
        cron_expression: {type: string}
        timezone: {type: [string, 'null']}
        enabled: {type: boolean}
        overlap_policy: {type: string, enum: [skip, queue, replace]}
    SchedulePatch:
      type: object
      minProperties: 1
      properties:
        cron_expression: {type: string}
        timezone: {type: [string, 'null']}
        enabled: {type: boolean}
        overlap_policy: {type: string, enum: [skip, queue, replace]}
    Schedule:
      allOf:
        - {$ref: '#/components/schemas/SchedulePut'}
        - type: object
          required: [id]
          properties:
            id: {type: string, format: uuid}
            next_run_at: {type: [string, 'null'], format: date-time}
            last_run_at: {type: [string, 'null'], format: date-time}
...